feat: add sidebar tab bar component
